requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==3.2
djangorestframework==3.12.4
h11==0.12.0
hiredis==2.0.0
idna==2.10
pytz==2021.1
redis==3.5.3
redisearch==2.0.0
requests==2.25.1
rmtest==0.7.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
uvicorn==0.13.4
